chore: doc integration command and cleanup
