+++ gno.land/pkg/integration/testscript_seqs.go
@@ -1,22 +1,1 @@
 package integration
-
-import (
-	"testing"
-
-	"github.com/rogpeppe/go-internal/testscript"
-)
-
-type tSeqShim struct{ *testing.T }
-
-// noop Parallel method allow us to run test sequentially
-func (tSeqShim) Parallel() {}
-
-func (t tSeqShim) Run(name string, f func(testscript.T)) {
-	t.T.Run(name, func(t *testing.T) {
-		f(tSeqShim{t})
-	})
-}
-
-func (t tSeqShim) Verbose() bool {
-	return testing.Verbose()
-}
